add ButtonBar with map
